refactor: getCurrentMember() HoneyTipPostFacade -> Controller로 이동, 누락 id 필드 변경점 추가

+++ src/main/java/com/api/ttoklip/domain/honeytip/service/HoneyTipImageService.java
@@ -1,10 +1,8 @@
 package com.api.ttoklip.domain.honeytip.service;
 
-import static com.api.ttoklip.global.util.SecurityUtil.getCurrentMember;
-
 import com.api.ttoklip.domain.honeytip.domain.HoneyTipImage;
-import com.api.ttoklip.domain.honeytip.repository.image.HoneyTipImageRepository;
 import com.api.ttoklip.domain.honeytip.domain.HoneyTip;
+import com.api.ttoklip.domain.honeytip.repository.image.HoneyTipImageRepository;
 import com.api.ttoklip.global.exception.ApiException;
 import com.api.ttoklip.global.exception.ErrorType;
 import java.util.List;
@@ -27,7 +25,7 @@ public class HoneyTipImageService {
     @Transactional
     public void deleteImages(final List<Long> imageIds, final Long currentMemberId) {
         validImagesExists(imageIds);
-        honeyTipImageRepository.allImageOwner(imageIds, currentMemberId);
+        honeyTipImageRepository.verifyMemberIsImageOwner(imageIds, currentMemberId);
         honeyTipImageRepository.deleteByImageIds(imageIds);
     }
